Fix cursor-decode livelock in kvstore.c; land accumulated correctness/perf work; repo hygiene
Critical fix: getAndClearHashtableIndexFromCursor() masked the scan
cursor with (num_hashtables-1) instead of (1<<num_hashtables_bits)-1.
That only coincides when num_hashtables is a power of two (true
upstream, e.g. 16384 cluster slots) -- NexCache's 176-shard kvstore
broke the assumption, corrupting every cursor round-trip (didx=89
decoded back as 9). kvstoreScan would land on the wrong empty shard,
skip it, correctly re-find the real one via the Fenwick tree, re-encode
it, and immediately mis-decode it again: an unbreakable ping-pong that
never reached real data. This made active-expire cycles spin at 400%+
CPU for minutes without ever expiring a key or firing its keyspace
notification -- reproduced reliably with as few as 2 keys in a db,
confirmed via live cursor tracing, fixed, and verified against the
real Redis TCL test suite (136/136 clean before an unrelated local
harness/port-reuse issue interrupted the run; CI will run it in full).

Also included (accumulated from prior sessions, never committed):
- SVI/RUBIN_MODE key-embedding fixes (SDS header type, INT-encoding
  read path) so SET/GET/KEYS/DBSIZE/INCR are correct under RUBIN_MODE.
- SET-path redesigned to invalidate-on-write + lazy read-through
  instead of write-through, closing VERAM3.3's SET-throughput gap.
- FLUSHALL/emptyDbAsync slot-count fix (SIGSEGV on flush+write burst).
- Parasitic legacy engine workers gated behind nex-engine-workers
  (off by default) -- was burning CPU with zero functional callers.
- FASE2 modules wiring vector/timeseries/streams to real RESP commands
  (VADD/VSEARCH, TS.*, XBP.*) plus ORSET_MEMBERS, previously
  write-only.
- Ported standalone concurrency/correctness test suite
  (tests/run_standalone_tests.sh, test_concurrency_20260807.c).

Repo hygiene: added a proper top-level .gitignore and untracked ~280
compiled build artifacts (.o/.a/.d/binaries) that had been committed
by mistake, causing spurious diffs on every local build. Removed two
leftover scratch debug executables. No functional changes from this
part.

+++ src/compression/auto.h
@@ -50,7 +50,15 @@ typedef enum DataType {
     DATATYPE_EMBEDDING = 5, /* Vettori float — poco comprimibili */
 } DataType;
 
-/* ── Header di un dato compresso (16 byte) ──────────────────── */
+/* ── Header di un dato compresso (19 byte) ────────────────────
+ * PRIMA: checksum era un XOR a 1 byte — ~1/256 di probabilità di
+ * accettare silenziosamente un payload corrotto in decompress_data, e
+ * trivialmente aggirabile da un flip di due byte con la stessa parità
+ * nella stessa posizione. Sufficiente come sanity check interno, non
+ * come integrity check se questo buffer attraversa un confine non
+ * fidato (replica, tiering su disco). Il campo è stato allargato a
+ * CRC32 (nessun dato reale è mai stato persistito con questo formato,
+ * dato che il modulo non è ancora collegato a un command path). */
 typedef struct __attribute__((packed)) CompressedHeader {
     uint32_t magic;           /* 0x4E584348 = "NXCH" */
     uint32_t original_size;   /* Dimesione dati originali */
@@ -58,7 +66,7 @@ typedef struct __attribute__((packed)) CompressedHeader {
     uint8_t comp_type;        /* CompressionType */
     uint8_t data_type;        /* DataType (hint) */
     uint8_t level;            /* Livello compressione */
-    uint8_t checksum;         /* XOR checksum dei byte dati */
+    uint32_t checksum;        /* CRC32 dei byte dati */
 } CompressedHeader;
 
 #define COMPRESS_MAGIC 0x4E584348U /* "NXCH" */
